+++ src/index.ts
@@ -173,6 +173,7 @@ export * from './services/finance-invoice-service';
 export * from './services/finance-payment-service';
 export * from './services/finance-construction-invoice-service';
 export * from './services/finance-construction-payment-service';
+export * from './services/finance-expense-service';
 export * from './services/finance-journal-service';
 export * from './services/finance-reports-service';
 export * from './services/finance-posting-account-service';